// lib/generated/permissions.ts
export const P = {
  ACCOUNT_CREATE: "account:create",
  ACCOUNT_READ: "account:read",
  ACCOUNT_UPDATE: "account:update",
  ACCOUNT_DELETE: "account:delete",
  ANALYTICS_READ: "analytics:read",
  BRANCH_CREATE: "branch:create",
  BRANCH_READ: "branch:read",
  BRANCH_UPDATE: "branch:update",
  BRANCH_UPDATE_PRODUCT_STOCK: "branch:update-product-stock",
  BRANCH_READ_PRODUCT_STOCK: "branch:read-product-stock",
  BRANCH_DELETE: "branch:delete",
  CUSTOMER_CREATE: "customer:create",
  CUSTOMER_READ: "customer:read",
  CUSTOMER_UPDATE: "customer:update",
  CUSTOMER_DELETE: "customer:delete",
  INVENTORY_RECEIPT_CREATE: "inventory-receipt:create",
  INVENTORY_RECEIPT_READ: "inventory-receipt:read",
  INVENTORY_RECEIPT_UPDATE: "inventory-receipt:update",
  INVENTORY_RECEIPT_VOID: "inventory-receipt:void",
  PAYMENT_CREATE: "payment:create",
  PAYMENT_READ: "payment:read",
  PAYMENT_VOID: "payment:void",
  PRODUCT_CREATE: "product:create",
  PRODUCT_READ: "product:read",
  PRODUCT_UPDATE: "product:update",
  PRODUCT_DELETE: "product:delete",
  SALE_CREATE: "sale:create",
  SALE_READ: "sale:read",
  SALE_UPDATE: "sale:update",
  SALE_VOID: "sale:void",
  SALES_SHIFT_READ: "sales-shift:read",
  SALES_SHIFT_ADD_EXPENSE: "sales-shift:add-expense",
  SALES_SHIFT_REMOVE_EXPENSE: "sales-shift:remove-expense",
  SALES_SHIFT_END_DAY: "sales-shift:end-day",
  SUPPLIER_CREATE: "supplier:create",
  SUPPLIER_READ: "supplier:read",
  SUPPLIER_UPDATE: "supplier:update",
  SUPPLIER_DELETE: "supplier:delete",
} as const;